chore: moved tests from test_utils to test_params, and well fixed a bunch of other things
