+++ tests/test_gradientboostedtrees.py
@@ -2,8 +2,10 @@ import numpy as np
 import pytest
 from sklearn.utils.estimator_checks import parametrize_with_checks
 
-import random_tree_models.decisiontree as dtree
 import random_tree_models.gradientboostedtrees as gbt
+from random_tree_models.decisiontree import (
+    DecisionTreeRegressor,
+)
 from tests.conftest import expected_failed_checks
 
 
@@ -36,9 +38,7 @@ class TestGradientBoostedTreesRegressor:
     def test_fit(self):
         model = gbt.GradientBoostedTreesRegressor()
         model.fit(self.X, self.y)
-        assert all(
-            [isinstance(model, dtree.DecisionTreeRegressor) for model in model.trees_]
-        )
+        assert all([isinstance(model, DecisionTreeRegressor) for model in model.trees_])
 
     def test_predict(self):
         model = gbt.GradientBoostedTreesRegressor()
@@ -67,9 +67,7 @@ class TestGradientBoostedTreesClassifier:
         model = gbt.GradientBoostedTreesClassifier()
         model.fit(self.X, self.y)
         assert not hasattr(self.model, "classes_")
-        assert all(
-            [isinstance(model, dtree.DecisionTreeRegressor) for model in model.trees_]
-        )
+        assert all([isinstance(model, DecisionTreeRegressor) for model in model.trees_])
 
     def test_predict(self):
         model = gbt.GradientBoostedTreesClassifier()
@@ -88,28 +86,3 @@ def test_gbt_estimators_with_sklearn_checks(estimator, check):
     Reference: https://scikit-learn.org/stable/modules/generated/sklearn.utils.estimator_checks.parametrize_with_checks.html#sklearn.utils.estimator_checks.parametrize_with_checks
     """
     check(estimator)
-
-
-@pytest.mark.parametrize(
-    "x,exp,is_bad",
-    [
-        (True, 1, False),
-        (False, -1, False),
-        ("a", None, True),
-        (1, 1, False),
-        (0, -1, False),
-        (-1, None, True),
-        (None, None, True),
-    ],
-)
-def test_bool_to_float(x, exp, is_bad: bool):
-    try:
-        # line to test
-        res = gbt.bool_to_float(x)
-    except ValueError as ex:
-        if is_bad:
-            pass  # Failed expectedly to convert non-bool values
-    else:
-        if is_bad:
-            pytest.fail(f"Passed unexpectedly for non-bool value {x} returning {res}")
-        assert res == exp
