+++ tests/test_leafweights.py
@@ -2,8 +2,7 @@ import numpy as np
 import pytest
 
 import random_tree_models.leafweights as leafweights
-import random_tree_models.params as utils
-from random_tree_models.params import MetricNames
+from random_tree_models.params import MetricNames, TreeGrowthParameters
 
 
 def test_leaf_weight_mean():
@@ -19,14 +18,14 @@ def test_leaf_weight_binary_classification_friedman2001():
 def test_leaf_weight_xgboost():
     g = np.array([1, 2, 3]) * 2
     h = np.array([1, 2, 3]) * 4
-    params = utils.TreeGrowthParameters(max_depth=2, lam=0.0)
+    params = TreeGrowthParameters(max_depth=2, lam=0.0)
     assert leafweights.leaf_weight_xgboost(g=g, h=h, growth_params=params) == -0.5
 
 
 class Test_calc_leaf_weight:
     def test_error_for_unknown_scheme(self):
         y = np.array([1, 2, 3])
-        growth_params = utils.TreeGrowthParameters(max_depth=2, lam=0.0)
+        growth_params = TreeGrowthParameters(max_depth=2, lam=0.0)
         with pytest.raises(NotImplementedError):
             leafweights.calc_leaf_weight(
                 y=y,
@@ -36,7 +35,7 @@ class Test_calc_leaf_weight:
 
     def test_leaf_weight_none_if_y_empty(self):
         y = np.array([])
-        growth_params = utils.TreeGrowthParameters(max_depth=2, lam=0.0)
+        growth_params = TreeGrowthParameters(max_depth=2, lam=0.0)
 
         weight = leafweights.calc_leaf_weight(
             y=y, growth_params=growth_params, measure_name=MetricNames.gini
@@ -45,7 +44,7 @@ class Test_calc_leaf_weight:
 
     def test_leaf_weight_float_if_y_not_empty(self):
         y = np.array([1, 2, 3])
-        growth_params = utils.TreeGrowthParameters(max_depth=2, lam=0.0)
+        growth_params = TreeGrowthParameters(max_depth=2, lam=0.0)
 
         weight = leafweights.calc_leaf_weight(
             y=y, growth_params=growth_params, measure_name=MetricNames.variance
