+++ tests/test_xgboost.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 from sklearn.utils.estimator_checks import parametrize_with_checks
 
-import random_tree_models.decisiontree as dtree
 import random_tree_models.xgboost as xgboost
+from random_tree_models.decisiontree import DecisionTreeRegressor
 from tests.conftest import expected_failed_checks
 
 
@@ -36,9 +36,7 @@ class TestXGBoostRegressor:
     def test_fit(self):
         model = xgboost.XGBoostRegressor()
         model.fit(self.X, self.y)
-        assert all(
-            [isinstance(model, dtree.DecisionTreeRegressor) for model in model.trees_]
-        )
+        assert all([isinstance(model, DecisionTreeRegressor) for model in model.trees_])
 
     def test_predict(self):
         model = xgboost.XGBoostRegressor()
@@ -67,9 +65,7 @@ class TestXGBoostClassifier:
         model = xgboost.XGBoostClassifier()
         model.fit(self.X, self.y)
         assert not hasattr(self.model, "classes_")
-        assert all(
-            [isinstance(model, dtree.DecisionTreeRegressor) for model in model.trees_]
-        )
+        assert all([isinstance(model, DecisionTreeRegressor) for model in model.trees_])
 
     def test_predict(self):
         model = xgboost.XGBoostClassifier()
